+++ node/src/types/kernel/api/system/logoutAuth/index.d.ts
@@ -17,11 +17,7 @@
 
 //#region content
 /* directories */
-export * as bootProgress from "./bootProgress";
-export * as currentTime from "./currentTime";
-export * as getConf from "./getConf";
-export * as logoutAuth from "./logoutAuth";
-export * as version from "./version";
 
 /* files */
+export * from "./response";
 //#endregion content